modify MemberService
modify (delete member, update member)

+++ gomin-sageori_Server/src/main/java/MAT/gominsageori/service/MemberService.java
@@ -2,20 +2,14 @@ package MAT.gominsageori.service;
 
 import MAT.gominsageori.domain.Member;
 import MAT.gominsageori.repository.MemberRepository;
-import MAT.gominsageori.repository.MemoryMemberRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import javax.transaction.Transactional;
-import java.util.List;
 import java.util.Optional;
 
-//@Service
+
 @Transactional
 public class MemberService {
     private final MemberRepository memberRepository;
 
-    //@Autowired
     public MemberService(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
     }
@@ -36,12 +30,17 @@ public class MemberService {
     public Optional<Member> findOne(String memberId) { return memberRepository.findById(memberId); }
 
     public String Update(Member member) {
-        memberRepository.findById(member.getId());
-        memberRepository.update(member);
-        return member.getId();
+        if(memberRepository.findById(member.getId()).isPresent()) {
+            memberRepository.update(member);
+            return member.getId();
+        }
+        else
+            return null;
     }
 
     public void delete(Member member) {
-        memberRepository.delete(member);
+        if(memberRepository.findById(member.getId()).isPresent()) {
+            memberRepository.delete(member);
+        }
     }
 }
